Added command groups package and RightRC
A command groups package was created. There is also an empty RightRC
command group to be filled.
