+++ src/org/usfirst/frc/team166/robot/commands/autonomous/Autonomous.java
@@ -2,7 +2,7 @@ package org.usfirst.frc.team166.robot.commands.autonomous;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
 
-import org.usfirst.frc.team166.robot.commands.lifts.InitLiftEncoders;
+import org.usfirst.frc.team166.robot.commands.groups.InitLiftEncoders;
 import org.usfirst.frc.team166.robot.commands.wings.RaiseLeftWing;
 import org.usfirst.frc.team166.robot.commands.wings.RaiseRightWing;
