+++ src/org/usfirst/frc/team166/robot/commands/groups/RightRC.java
@@ -1,0 +1,13 @@
+package org.usfirst.frc.team166.robot.commands.groups;
+
+import edu.wpi.first.wpilibj.command.CommandGroup;
+
+/**
+ *
+ */
+public class RightRC extends CommandGroup {
+
+	public RightRC() {
+
+	}
+}
